modifier un jour férié ok

+++ src/main/java/com/repository/PublicHolidayRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 
 import java.time.LocalDate;
 
-public interface PublicHolidayRepository  extends JpaRepository<PublicHolidays, Integer> {
+public interface PublicHolidayRepository  extends JpaRepository<PublicHolidays, Long> {
     public PublicHolidays findByDate(LocalDate date);
 
     @Query("SELECT ph FROM PublicHolidays ph WHERE YEAR(date) = :year ORDER BY date ASC")
